pyproject.toml
[tool.poetry]
name = "pythonsul2022"
version = "0.1.0"
description = ""
authors = ["Francisco Alface <falface@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pelican = "4.2.0"
tornado = "6.0.3"
docutils = "0.15.2"
invoke = "1.3.0"
livereload = "2.6.1"
python-decouple = "3.3"
MarkupSafe = "1.1.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.3.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
